+++ code/src/cljs/sixsq/nuvla/ui/welcome/views.cljs
@@ -30,7 +30,6 @@
 (defmethod panel/render :welcome
   [path]
   (let [tr           (subscribe [::i18n-subs/tr])
-        iframe?      (subscribe [::main-subs/iframe?])
         query-params (subscribe [::main-subs/nav-query-params])
         {:keys [message, error]} @query-params]
     (when @query-params